requirements.txt
Flask==0.7.2
python-firebase==1.2